pyproject.toml
[project]
name = "tech-challenge-four"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.42.1",
    "duckdb>=1.4.2",
    "fastapi>=0.121.3",
    "matplotlib>=3.10.7",
    "pandas-ta>=0.3.14b",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
